Exit status correct
